condense path logic, move text_field function to utils
