+++ backend/sparrow/api/filters.py
@@ -5,7 +5,7 @@ import datetime
 
 from sparrow.context import app_context
 from ..datasheet.utils import create_bound_shape
-from .utils import join_path, join_loops
+from .utils import join_path, join_loops, text_fields
 
 
 def _schema_fields(schema):
@@ -19,15 +19,6 @@ def create_params(description, example):
 
 def create_model_name_string(model):
     return f"{model}".split(".")[-1].split("'")[0].lower()
-
-
-def has_join_path(start, end):
-    """
-    Function to be used in the Should_Apply for filters.
-    """
-    path = join_path(start, end)
-    return path is not None
-
 
 class BaseFilter:
     params = None
@@ -176,7 +167,7 @@ class DateFilter(BaseFilter):
         answer = (
             hasattr(self.model, "date")
             or hasattr(self.model, "session_collection")
-            or has_join_path(model_name, "session")
+            or join_path(model_name, "session") is not None
         )
         return answer
 
@@ -197,8 +188,8 @@ class DateFilter(BaseFilter):
         if hasattr(self.model, "session_collection"):
             return query.join(self.model.session_collection).filter(and_(Session.date > start, Session.date < end))
 
-        elif has_join_path(model_name, "session") and len(join_path(model_name, "session")) > 1:
-            path = join_path(model_name, "session")
+        path = join_path(model_name, "session")
+        if path is not None and len(path) > 1:
             db_query = join_loops(path, query, db, self.model)
 
             return db_query.filter(and_(Session.date > start, Session.date < end))
@@ -221,7 +212,7 @@ class DOIFilter(BaseFilter):
         answer = (
             hasattr(self.model, "doi")
             or hasattr(self.model, "publication_collection")
-            or has_join_path(model_name, "publication")
+            or join_path(model_name, "publication") is not None
         )
         return answer
 
@@ -238,8 +229,8 @@ class DOIFilter(BaseFilter):
         if hasattr(self.model, "publication_collection"):
             return query.join(self.model.publication_collection).filter(publication.doi.like(f"%{doi_string}%"))
 
-        if has_join_path(model_name, "publication") and len(join_path(model_name, "publication")) > 1:
-            path = join_path(model_name, "publication")
+        path = join_path(model_name, "publication")
+        if path is not None and len(path) > 1:
             db_query = join_loops(path, query, db, self.model)
 
             return db_query.filter(publication.doi.like(f"%{doi_string}%"))
@@ -260,7 +251,7 @@ class CoordinateFilter(BaseFilter):
 
     def should_apply(self):
         model_name = create_model_name_string(self.model)
-        return hasattr(self.model, "location") or has_join_path(model_name, "sample")
+        return hasattr(self.model, "location") or join_path(model_name, "sample") is not None
 
     def apply(self, args, query):
         if self.key not in args:
@@ -273,8 +264,8 @@ class CoordinateFilter(BaseFilter):
         pnts = [int(pnt) for pnt in points]
         bounding_shape = create_bound_shape(pnts)
 
-        if has_join_path(model_name, "sample") and len(join_path(model_name, "sample")) > 1:
-            path = join_path(model_name, "sample")
+        path = join_path(model_name, "sample")
+        if path is not None and len(path) > 1:
             db_query = join_loops(path, query, db, self.model)
             sample = db.model.sample
 
@@ -299,7 +290,7 @@ class GeometryFilter(BaseFilter):
         return (
             hasattr(self.model, "location")
             or hasattr(self.model, "sample_collection")
-            or has_join_path(model_name, "sample")
+            or join_path(model_name, "sample") is not None
         )
 
     def apply(self, args, query):
@@ -317,9 +308,9 @@ class GeometryFilter(BaseFilter):
             return query.join(self.model.sample_collection).filter(
                 func.ST_GeomFromEWKT(WKT_query).ST_Contains(func.ST_Transform(sample.location, 4326))
             )
-
-        elif has_join_path(model_name, "sample") and len(join_path(model_name, "sample")) > 1:
-            path = join_path(model_name, "sample")
+        
+        path = join_path(model_name, "sample")
+        if path is not None and len(path) > 1:
             db_query = join_loops(path, query, db, self.model)
 
             return db_query.filter(
@@ -343,7 +334,7 @@ class AgeRangeFilter(BaseFilter):
 
     def should_apply(self):
         model_name = create_model_name_string(self.model)
-        return model_name == "datum" or hasattr(self.model, "datum_collection") or has_join_path(model_name, "datum")
+        return model_name == "datum" or hasattr(self.model, "datum_collection") or join_path(model_name, "datum") is not None
 
     def apply(self, args, query):
         if self.key not in args:
@@ -357,8 +348,9 @@ class AgeRangeFilter(BaseFilter):
         if hasattr(self.model, "datum_collection"):
             datum = db.model.datum
             return query.join(self.model.datum_collection).filter(datum.value < age)
-        elif has_join_path(model_name, "datum") and len(join_path(model_name, "datum")) > 1:
-            path = join_path(model_name, "datum")
+
+        path = join_path(model_name, "datum")
+        if path is not None and len(path) > 1:
             db_query = join_loops(path, query, db, self.model)
             datum = db.model.datum
 
@@ -429,24 +421,6 @@ class IdListFilter(BaseFilter):
         return query.filter(self.model.id.in_(ids))
 
 
-## TODO: Age range filter, open parameter i.e 'Plateua Ages', pass a number and operator?
+## TODO: Open parameter i.e 'Plateua Ages', pass a number and operator?
 ## TODO: Define filter in plugin, i.e irradiation filter for WiscAr
 ## TODO: Filter based on nested models, /api/v2/datum?nest=project,datum_type&datum_type.unit=Ma&project=11
-
-
-def text_fields(model):
-    """
-    Function to return the column model attributes for a sqlalchemy model whose type is text
-    i.e sample.name
-    """
-    fields = model.__table__.columns.keys()
-    text_fields = []
-    for c in fields:
-        if f"{getattr(model,c).type}" == "TEXT":
-            text_fields.append(c)
-
-    atr = []
-    for c in text_fields:
-        atr.append(getattr(model, c))
-
-    return atr
